requirements.txt
python-telegram-bot==20.0
openai
apscheduler
python-dotenv
ephem
pytz